Changes to compile with latest version of g++.
TEMPLATE_SPECIALIZATION() macro updated; added
compiler define NEWSTYLE_TEMPLATES.  Formerly,
the alternate "new standard" way of declaring
template specializations only applied to BCC 5.5.
For older compilers, should not define NEWSTYLE_TEMPLATES.
It is defined by default in the generic makefile.

-fguiding-decls removed from command line for
g++; no longer supported.

Most other changes involve silencing warnings

+++ sources/gui/efglayout.cc
@@ -34,7 +34,7 @@ inline void DrawLine(wxDC &dc, double x_s, double y_s, double x_e, double y_e,
                      const wxColour &color, int thick = 0)
 {
   dc.SetPen(wxPen(color, (thick) ? 8 : 2, wxSOLID));
-  dc.DrawLine(x_s, y_s, x_e, y_e);
+  dc.DrawLine((int) x_s, (int) y_s, (int) x_e, (int) y_e);
 }
 
 inline void DrawRectangle(wxDC &dc, int x_s, int y_s, int w, int h,
